change marketId type from uint8 to uint16

+++ client/http_market_info.go
@@ -6,24 +6,24 @@ import (
 
 // OrderBook represents a market on Lighter exchange
 type OrderBook struct {
-	MarketID                uint8   `json:"market_id"`
-	Symbol                  string  `json:"symbol"`
-	Status                  string  `json:"status"`
-	TakerFee                string  `json:"taker_fee"`
-	MakerFee                string  `json:"maker_fee"`
-	LiquidationFee          string  `json:"liquidation_fee"`
-	MinBaseAmount           string  `json:"min_base_amount"`
-	MinQuoteAmount          string  `json:"min_quote_amount"`
-	SupportedSizeDecimals   int     `json:"supported_size_decimals"`
-	SupportedPriceDecimals  int     `json:"supported_price_decimals"`
-	SupportedQuoteDecimals  int     `json:"supported_quote_decimals"`
+	MarketID               uint16 `json:"market_id"`
+	Symbol                 string `json:"symbol"`
+	Status                 string `json:"status"`
+	TakerFee               string `json:"taker_fee"`
+	MakerFee               string `json:"maker_fee"`
+	LiquidationFee         string `json:"liquidation_fee"`
+	MinBaseAmount          string `json:"min_base_amount"`
+	MinQuoteAmount         string `json:"min_quote_amount"`
+	SupportedSizeDecimals  int    `json:"supported_size_decimals"`
+	SupportedPriceDecimals int    `json:"supported_price_decimals"`
+	SupportedQuoteDecimals int    `json:"supported_quote_decimals"`
 }
 
 // OrderBooksResponse represents the response from order_books API
 type OrderBooksResponse struct {
-	Code        int          `json:"code"`
-	Message     string       `json:"message"`
-	OrderBooks  []OrderBook  `json:"order_books"`
+	Code       int         `json:"code"`
+	Message    string      `json:"message"`
+	OrderBooks []OrderBook `json:"order_books"`
 }
 
 // GetOrderBooks fetches all available markets from Lighter exchange
@@ -37,18 +37,18 @@ func (c *HTTPClient) GetOrderBooks() (*OrderBooksResponse, error) {
 }
 
 // GetMarketByID fetches information for a specific market
-func (c *HTTPClient) GetMarketByID(marketID uint8) (*OrderBook, error) {
+func (c *HTTPClient) GetMarketByID(marketID uint16) (*OrderBook, error) {
 	orderBooks, err := c.GetOrderBooks()
 	if err != nil {
 		return nil, err
 	}
-	
+
 	for _, ob := range orderBooks.OrderBooks {
 		if ob.MarketID == marketID {
 			return &ob, nil
 		}
 	}
-	
+
 	return nil, fmt.Errorf("market ID %d not found", marketID)
 }
 
@@ -58,13 +58,13 @@ func (c *HTTPClient) GetMarketBySymbol(symbol string) (*OrderBook, error) {
 	if err != nil {
 		return nil, err
 	}
-	
+
 	for _, ob := range orderBooks.OrderBooks {
 		if ob.Symbol == symbol {
 			return &ob, nil
 		}
 	}
-	
+
 	return nil, fmt.Errorf("market symbol %s not found", symbol)
 }
 
@@ -74,14 +74,14 @@ func (c *HTTPClient) GetActiveMarkets() ([]OrderBook, error) {
 	if err != nil {
 		return nil, err
 	}
-	
+
 	activeMarkets := []OrderBook{}
 	for _, ob := range orderBooks.OrderBooks {
 		if ob.Status == "active" {
 			activeMarkets = append(activeMarkets, ob)
 		}
 	}
-	
+
 	return activeMarkets, nil
 }
 
@@ -91,10 +91,10 @@ func (c *HTTPClient) PrintMarkets() error {
 	if err != nil {
 		return err
 	}
-	
+
 	fmt.Println("Lighter Markets:")
 	fmt.Println("================")
-	
+
 	for _, ob := range orderBooks.OrderBooks {
 		fmt.Printf("Market ID: %d\n", ob.MarketID)
 		fmt.Printf("Symbol: %s\n", ob.Symbol)
@@ -105,6 +105,6 @@ func (c *HTTPClient) PrintMarkets() error {
 		fmt.Printf("Min Quote Amount: %s\n", ob.MinQuoteAmount)
 		fmt.Println("----------------")
 	}
-	
+
 	return nil
 }
